requirements.txt
# Django==5.2.1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
redis==5.3.0
celery==5.3.1
